move specs to matching screen directory
